Fixed a bug in dbms_alert_defered_signal: if message parameter was NULL, we passed the NULL to DatumGetTextP causing a segfault.
Fixed a bug in unregister_event: failed to unregister the last receiver
on an event, because we incorrectly tested for
"ev->receivers_number > 1", instead of "ev->receivers_number > 0"

Small improvements in error messages: "good event/relatio" (sic) ->
"valid event/relation". There's still a lot of room for improvement, but
it's a start.