Moved pg_typeof() to a patch.
The pg_typeof() function is included in 8.4, so there was code removing it
from `pgtap.sql` when building on 8.4 and later. I'd rather have the latest
version of PostgreSQL make no changes to the file, so I've changed things
around. Now `pg_typeof()` is no longer in `pgtap.sql.in`, but is added by
`compat/install-8.3.patch.`. This keeps things clean (and line numbers
correct!) when I'm doing development on 8.4.