Reorganized bag and set function documentation.
As I was writing a blog entry for 0.22, I realized that it made more sense to
keep the set and bag functions grouped spearately: sets with sets and bags
with bags.

I also updated the link to the summary of result set function arguments
to the new name of the section.