Add `operators_are()`.
Also changed `casts_are()` to accept an array of strings rather than a
multidimensional array of types. It's just easier, and close enough to the
syntax of `CREATE CAST` to make sense. Did this after I decided that a similar
approach to `operators_are()` made much more sense, given that `::regoperator`
nicely formats operators.

In the process, added `_areni()`, which is the same as `_are()` except that it
doesn't `quote_ident()` extra or missing things. Used for the string syntax of
`casts_are()` and `operators_are()`.