Fix `is_member_of()` and improve 8.0 Compat.
Turns out that the way that groups are represented in 8.1 and higher is just a
simulation of how they looked before 8.1. This is because roles are different,
of course. So I changed the query in `_grolist()` to use the proper role way
of doing things, and then took the old query and added it to the 8.0 patch.

Of course, that required a lot of work, so I went ahead and changed how the
8.0 patch works so that it doesn't replicate the work of other patches. To do
this, I had to move the code that removes `E''` syntax to run *after* the
patches have been applied. This is a lot cleaner, and should make maintenance
of the 8.0 compat patch a lot simpler from here on in. Change tested with
8.3-8.4.