No more PKs on internal tables.
There was code in `plan()` and `_cleanup()` to temporarily change
`client_min_messages` to "warning" as a way to suppress NOTICEs when creating
and dropping the tables that pgTAP uses internally. By eliminating the primary
keys in those tables in favor of `NOT NULL` columns and `UNIQUE INDEX`es,
those workarounds are no longer necessary.