Fix to support $TAPSCHEMA again.
The patch management is such a PITA. Anyway, needed to not change the files
until after the patches are applied, as doing so can break the patches.

While at it, remove the extra diff files from `.gitignore`, as I actually want
to know when they're there. One was mis-spelled anyway. Go figure.