Nicer handling of IDs in HTML documtation.
Update `Makefile` to create nicer IDs for function headers. So instead of an
ID of "`isnt(+anyelement,+anyelement,+description+)`" we'll get an ID of
"isnt". This will make for much nicer links into the documentation.

This required that I change the IDs generated by Discount. I think this is a
good thing.

While at it, fix some broken backticks in `README.pgtap` that were messing up
the HTML output a bit.