Improve handling of expressions in `has_index()`.
Rather than allowing only one expression, allow columns and one expression to
be mixed in a single query. This is much closer to what one should expect. The
downside is that, if an index has multiple expressions, there is no way to
separate them, yet. Will fix if I get an answer to post to -hackers about it.
Meanwhile, this should get us further along than we were before, and requires
less code, to boot! (Ref #12.)