Add fake tests for foreign_tables_are functions in older postgres.
Older version of postgres do not support foreign_tables. Rewrite
the tests for the foreign_tables_are functions to check the version
and run the real tests if foreign tables are supported. If an older
version of postgres run pass/fail tests which create the same output
as the real tests.