Estimated average row size check for btree indexes
If the estimated average row size >= max btree item size, return an error.

If it's >= 80% of max btree item size, just warn a the real index would probably
fail. Since it's an estimated and average row size, 80% if probably really not
aggressive enough, but still better than nothing. A GUC for this "security area"
could be useful.