Fixed #6: Audit log is not output after the SET ROLE.
This issue had to do with trying to suppress audit logs for statements that are not run as the current role, which includes security definer functions and foreign key lookups.  If the initial logon was a superuser then no audit logs would be created at all after a set role and while this is not very realistic it still needs to be taken in account.  The downside of this change is that FK lookups and statements in security definer functions will be logged when the user initially logs on a superuser making the audit log much noisier in this case.