Setup YB arcanist
Summary:
Setting up base .arcconfig so we can run `arc diff`. You'll need to
run `arc install-certificate` so that your local arcanist has a token to talk
to the remote Phabricator.

You also need to add your public SSH key to Phabricator so that you are
recognized.

Test Plan: this diff is up

Reviewers: mikhail, kannan

Reviewed By: kannan

Differential Revision: http://phabricator.dev.yugabyte.com:80/D3