ENG-61: CPU profiling relies on non-existent directory
Summary:
CPU profiler handler was assuming the presence of a directory /tmp/yb_cpu_profile, and trying to use a file name of the form:
   /tmp/yb_cpu_profile/$0.$1
The code isn't creating the directory if it doesn't exist. Fixing this by changing the file format to:
  /tmp/yb_cpu_profile.$0.$1
and not relying on the sub-directory being present. The heap profiler already just puts its files in /tmp.

Test Plan:
Manually tested by invoking a http request against the tserver:

/pprof/profile?seconds=1
/pprof/heap?seconds=2

and verified the log lines, and the profile output.

Reviewers: bogdan

Reviewed By: bogdan

Subscribers: bharat, karthik, mikhail

Differential Revision: http://phabricator.dev.yugabyte.com:80/D6