ENG-64: Expose num_tablet_servers_live as a metric
Summary: num_tablet_servers_live reports the number of tservers that have responded within the last 60 seconds. Note that only the master leader reports the correct number and others report 0.

Test Plan:
Ran this on the local cluster and verified that the correct number was reported, also verified that this was reduced to 2 when one tserver was left killed for sometime.
    {
        "type": "cluster",
        "id": "yb.cluster",
        "attributes": {},
        "metrics": [
            {
                "name": "num_tablet_servers_live",
                "value": 0
            }
        ]
    }

Reviewers: mikhail, kannan

Reviewed By: kannan

Subscribers: jenkins-bot, eng

Differential Revision: http://phabricator.dev.yugabyte.com:80/D9