Switching mechanics of creating a cluster
Summary:
Added new flag for creating a cluster: create_cluster, which must be used in conjunction with the master_addresses flag.
The create_cluster flag should only ever be set the first time you setup the cluster. Every restart should not have the flag set.
If you are asked to create a cluster and there is local cmeta data, the startup will fail!

Right now, the mechanics will be as follows:
- create_cluster=true, master_addresses.empty()=true: Create a new cluster in non-distributed mode. Fail if cmeta data exists.
- create_cluster=true, master_addresses.empty()=false: Create a new cluster in distributed mode with the listed masters. Fail if cmeta data exists.
- create_cluster=false, master_addresses.empty()=true: Assumed non-distributed restart, will fail if local cmeta does not exist.
- create_cluster=false, master_addresses.empty()=false: Assumed distributed restart, will fail if local cmeta does not exist. This should be replaced with mechanics for an empty master.

Test Plan: Tested with local_cluster script and manual `rm -rf` for the various cases.

Reviewers: karthik, bharat

Reviewed By: bharat

Subscribers: kannan, mikhail, jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D12