ENG-81 Try to fix RemoteBootstrapITest.TestDeleteLeaderDuringRemoteBootstrapStressTest by bumping up the number of rows inserted at each iteration
Summary: RemoteBootstrapITest.TestDeleteLeaderDuringRemoteBootstrapStressTest waits for bootstrap after killing a tablet server. However, it looks like the bootstrap may happen too quickly and we may not see a "COPYING" state, but see a "READY" state right away. In this case the test times out waiting for the "COPYING" state. One potential way to fix it is to bump up the number of rows inserted at each iteration of the test to make the bootstrap slower.

Test Plan: Run the test multiple times

Reviewers: bogdan, bharat

Reviewed By: bharat

Subscribers: jenkins-bot, karthik

Differential Revision: http://phabricator.dev.yugabyte.com:80/D17