ENG-68 Use Raft indexes as RocksDB sequence numbers
Summary:
Modifying RocksDB's WriteBatch to make it possible to specify sequence numbers for each update. The user-specified sequence numbers are stored in a separate vector field inside WriteBatch and not in the binary-serialized `rep_` string to make the RocksDB change less invasive and make it easier to cherry-pick upstream improvements in the future. Also, for the same reason, the API to set user-defined sequence numbers is a separate method call, `AddUserSequenceNumber`, that has to be called before methods adding updates to a WriteBatch are called, and not a new parameter to `Put`, `Delete`, `Merge` methods.

The user-defined sequence number vector inside WriteBatch either has to be empty or be of the same length as the value of `Count()` (the number of updates in the WriteBatch). An empty batch is assumed not to have any user-defined sequence numbers. Only batches with the same value of `HasUserSequenceNumbers()` are allowed to participate in an append operation and be used in the same write group in `DBImpl::WriteImpl`.

Build improvements:
* Added a precompiled header file for RocksDB. This improved the build by 25% in my benchmark.
* Unifying the `--build-all-targets` and `--target` options in the `build-rocksdb.sh` script into a single `--targets` option. The special target "all" now corresponds to what `--build-all-targets` used to do (it does not map directly to the "all" target in the RocksDB Makefile).
* `yb_build.sh` asks the user for confirmation when doing a clean build if another clean build was done less than an hour ago.
* It is now possible to build a specific RocksDB test with `yb_build.sh` by specifying the test name prefixed by `rocksdb_` on the command line.

Test Plan:
* All existing YB and RocksDB unit tests
* New unit tests: `user_sequence_number_test` and a new test in `write_batch_test`
* Added a small multi-threaded insertion test to `kv_table-test.cc`
* Manual testing using yb_load_test_tool

Reviewers: karthik, kannan

Reviewed By: kannan

Subscribers: kannan

Differential Revision: http://phabricator.dev.yugabyte.com:80/D8