Making linter singlethread
Summary:
The git-clang-format needs to execute a git command, which needs an
actual git lock. I did not realize this while setting up the Lint class using
a baseclass that allowed for parallelization.

I've modified the code now to be simple and straightforward. I also realized
that it should be relatively easy to make it both parallel and automatically
detecting the merge-base so we don't need the hardcoded HEAD^.

Test Plan: ran on one of my changes.

Reviewers: karthik, bharat, mikhail, kannan

Reviewed By: kannan

Subscribers: jenkins-bot

Differential Revision: http://phabricator.dev.yugabyte.com:80/D21