ENG-83: Fixing the HEAD^ hack on linting
Summary:
Figured out the hooks in the lint engine that we can use to get the
commit hash that is the merge-base to compare against. I'm just comparing
against master now, but we could have it compare the tracked branch, if needed.

Test Plan: modified 2 files, confirmed linter worked as expected

Reviewers: mikhail, kannan

Reviewed By: kannan

Differential Revision: http://phabricator.dev.yugabyte.com:80/D23