ENG-47: Per tablet server RaftConfig disappears from UI on master addition
Summary:
Found the root cause that the master leader was reloading the sysCatalog after a master addition, erasing the tablet replica info used in the UI. This is because the SysCatalogStateChanged() API was assuming that if the current role is RaftPeerPB::LEADER, it should issue a DoElectionCallback(). But for master addition case, the LEADER_CONFIG_CHANGE_COMPLETE should not trigger this callback, as LEADER did not change role...

There was a second load of metadata as an actual leader election was kicking in - this was more of a side-effect of starting the new master in non-shell mode. The new master was moving its term ahead while pinging the existing cluster, and the leader sees the term is ahead of its own, steps down and then ends up starting and winning the election (as no one else votes for new master). For now, if the new master is started with --enable_leader_failure_detection=false it solves the issue.

Test Plan:
Manually tested using yb-admin ADD_SERVER that the UI shows the raft config correctly after master addition and ensured that the unexpected metadata reloads are not called.

Reviewers: bogdan, karthik

Reviewed By: karthik

Subscribers: jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D32