Adding redirects to Master Leader across different Master UI tabs
Summary:
Adding redirect link for Master UIs across views, to link to the Master Leader.
This is useful as the Leader actually has the in-memory state that we need, to
for example, view the tablets assigned to various tablet servers.

Test Plan:
Quick screenshot of the view on a non-leader:
https://phabricator.dev.yugabyte.com/F81 (notice the link at the bottom)

It will also relay any query parameters as well, if we have pages that need
that.

Reviewers: karthik, ram, bharat

Reviewed By: bharat

Subscribers: jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D43