Adding mechanics to balance tablet distribution in real time
Summary:
Enhanced the mechanics of CatalogManager to also contain a LoadBalancer. This will allow it to shift load across the cluster, if the variance in load between machines becomes too high (for example, a new node is added, which has 0 tablets).

I created a separate class, to keep the mechanics separate, but right now it is a friend class of CatalogManager, which is something undesirable. Ideally, we should figure out which state it should have access to and enable external access.

The algorithm right now prioritizes passing load to the least loaded servers, rather than always looking for highest variance pairs. It has mechanics to prevent having too many tablets be over-replicated, too many starting across the cluster and too little load variance between servers (no point moving a tablet if the servers differ in load by 1).

Test Plan: Tested locally. I also added a flag to enable/disable the load balancing, which defaults to disabled. In a follow-up diff, I will switch it to enabled and fix the tests that make the implicit assumption that tablets should not be moving.

Reviewers: kannan, bharat, karthik, mikhail

Reviewed By: mikhail

Subscribers: mikhail, jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D55