Fix test ordering issue in do_tap.sql.
The tap-emitting functions were ordering their output by the text of the tap, which
starts with "# $test_number", but the new tests added in 83cb5191 caused the numbers
to cross from single-digit to two-digit, so "# 10" was sorting before "# 9". So
add another test that finds no tests, which is useful anyway, and pushes all
the test result counts above two digits.