ENG-118: Illegal state: Latest committed opid 0.0 is different from current term 1
Summary:
Some tests intermittently fail the above exception. Turns out, the tests use the WaitForServersToAgree() as way of making sure all servers have at least gotten up to a specified "minimum_index", and that their indexes are the same.

  ASSERT_OK(itest::WaitForServersToAgree(timeout, ts_map_, tablet_id, 1 /* minimum_index */)); // Wait for NO_OP.

When we return from WaitForServersToAgree() we can only be assured that the last opid in each replica is the same and at least "minimum_index". This doesn't necessarily mean that the leader knows about this. So as far as the leader is concerned the "committed_op_index" may not have yet been advanced since these peers might not yet have responded back to the leader.

If meanwhile, the leader gets a racing request for "ChangeConfig" (in this case, the test was doing a RemoveServer() which would result in a ChangeConfig()). The request was coming right after the WaitForServersToAgree() call in the test. So, what's happening is that even though each replica's last opid is 1.1, the leader still thinks that the last committed op is 0.0 (as he has not yet heard back from all the replicas). And so when the ChangeConfig() request comes in, leader rejects it because of the requirement that we should have at least one operation in our current term as leader before allowing a config change. And since, as far as the leader is concerned, there is no index committed in term 1 yet, the ChangeConfig operation is reject with an "Illegal State" error.

(https://phabricator.dev.yugabyte.com/diffusion/YB/browse/master/src/yb/consensus/raft_consensus.cc;9092e8cd4b23372fcf08c932de35efe0d44c3a3f$1382-1385).

Changing the test(s) to wait for the leader's commit index to reach expected value addresses the issue.

Test Plan:
I was able to reproduce the problem at will if I introduce a small delay in UpdateMajorityReplicated(); and then confirm that the fix solves the issue.

Reviewers: bharat

Reviewed By: bharat

Subscribers: jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D65