Fixing RPC timeout issue with poll
Summary:
We discovered a weird issue with RPCs timing out on starting the
cluster locally on mac. Apprently, the call to poll/ppoll takes a timeout
parameter, which for Mac is an int, but is being passed in a long. This seemed
to magically work until now, probably due to truncation luck.

The intended desire is to set a timeout on the RPC controller, which most
callsites do, except in this case, when we get the node registration for the
masters on startup, where no timeout is set. This defaults the deadline
parameter to MonoTime::Max in src/yb/rpc/reactor.cc:230. This in turn leads to
the eventual overflow in the negotiation call.

Test Plan:
Re-ran locally, works. Should confirm that the respective change to
Linux codepath also works.

Reviewers: bharat, kannan, mikhail

Reviewed By: kannan

Subscribers: jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D67