Updates error diagnostics output for 9.2.
The schema, table, column, data type, and constraint contexts were not
included until 9.3, so add a patch to remove them and update the tests to
generate and test for their presence programmatically. This requires
additional expected output files for a couple of the tests, since the output
varies. Fortunately, pg_regress supports up to 9 alternate output files.

I've also added alternate output files for those tests for 9.4, where there
was a very slight change in the output of context in 9.5 to omit the word
"statement" from the first line of context.