Update for 8.4.
Includes a new patch to remove the `DROP FUNCTION file has_exceptions()
statements from the uninstall file. Speaking of which, we add those statemetns
to the uninstall file and update it with other stuff that has changed since
who knows when. As a result, drop a duplicate function declarationfor
_get_schema_privs() from the install file and make the layout of the
performs_within() and _error_diag() functions the same as all the others, so
it's easy to regenerate the uninstall file.