Refactor of Registration information in preparation of multi-az support
Summary:
Need to join the TS and Master registration info in preparation of
AZ awareness. I did some refactoring and added the base ServerRegistration as a
common field in TSRegistration. If we ever need a special one for Master, we
can do the same.

Test Plan: running tests on Jenkins. Cluster ran locally with load test.

Reviewers: karthik, mikhail, bharat

Reviewed By: bharat

Subscribers: jenkins-bot

Differential Revision: https://phabricator.dev.yugabyte.com/D80