[ENG-146] YBAdmin list-tables command also shows key-ranges
Summary:
the key range information is added to the tablet descriptions

Comment: there seems to be an existing bug that all the tablets are not shown when running this command. When using default load-tester, there should be 16 tablets but 10 are shown. I'll look into the issue and update the diff later to fix it.

Test Plan: Tested by running yb-admin

Reviewers: kannan, bharat

Reviewed By: kannan, bharat

Subscribers: mikhail, kannan, jenkins-bot, karthik, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D83