ENG-160: Adding mechanics to stop load balancing
Summary:
Added methods on the masters to change the state of load balancing.
This is currently not persistent, until we decide on a format to persist
settings, either in sys catalog, or in a new sys table, etc.

Test Plan:
Restarted local cluster, issued command, checked all masters that
they got the command. Confirmed leader stopped doing load balancing.

Reviewers: karthik, bharat

Reviewed By: bharat

Subscribers: mikhail, jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D99