Refactoring PlacementInfo to have a CloudInfo proto for reuse
Summary:
Splitting diffs to do PlacementInfo refactoring as a standalone
change. We plan to have cluster level information and table level information,
but those will be two different diffs and we need the foundational PBs to be
already there.

Test Plan:
running tests on Jenkins. Since this is mostly a refactor, there
shouldn't be any issues...

Reviewers: karthik, bharat

Reviewed By: bharat

Subscribers: eng, jenkins-bot

Differential Revision: https://phabricator.dev.yugabyte.com/D114