ENG-179: Delete on-disk state of a removed master which got moved to shell mode
Summary:
Added calls to existing API which remove the wal, cmeta and tablet superblocks during going to shell mode of a removed master.

Test Plan: Manually tested that a removed master in shell mode can be added back to the cluster - it doesnt use any stale metadata, and ends up doing a remote bootstrap of the system tablet.

Reviewers: kannan, bogdan

Reviewed By: bogdan

Subscribers: bogdan, jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D125