Adding API to change cluster config
Summary:
Added API for changing cluster config across service, catalog manager and sys_catalog.
Also added mechanics to support client-side read-modify-write, so that we can have a single,
unified interface for changing the configs, as opposed to writing new APIs on the server everytime
we wish to have a custom update for a certain field combination from the config.

Test Plan: added test for sys_catalog and catalog manager API.

Reviewers: kannan, bharat

Reviewed By: kannan, bharat

Subscribers: jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D138