Glue for cluster and table level placement info
Summary:
Now that we have the API to modify cluster level config, we had to glue together the
custer level config change with the table level one. Adding changes in the client so that we can
also test this at the cluster level and port these changes to Java for the middleware.

Added extra test for create table to confirm these work as expected.

Test Plan: test runs as expected

Reviewers: kannan, karthik, bharat

Reviewed By: bharat

Subscribers: jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D142