ENG-154: Making load balancer AZ aware
Summary:
Refactored PlacementInfoPB to be a self-contained proto with a repeated field inside
for PlacementBlockPB. Used that to enhance load balancer to take into account placement info on
tablet moves.

For adding replicas, it will not move tablets to AZs that do not match the placements.
For removing replicas:
- it will explicitly remove replicas from over-replicated tablets, that live in AZs that are not
part of the placement anymore (for config changes).
- if all replicas match the placements, then it will consider the specified minimum number of
replicas, and remove over-replicated ones from those that have more than said minimum.

Test Plan: added tests for this.

Reviewers: kannan, karthik, bharat

Reviewed By: bharat

Subscribers: mikhail, jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D153