Porting the placement and blacklist mechanics to java client
Summary:
Added java pieces for request and response for changing the cluster
config. This allows us to soon use this API from the middleware and be able to
set and change the placement and use the blacklist to drain load from machines
we are moving off of.

Test Plan:
Tested using the cli. Started a cluster on 3 machines. Started load
tester and confirmed equal load. Added a 4th machine and confirmed load balance
between all of them. Blacklisted one of the machines and confirmed load is
drained, except for the tablets that have the leaders on that blacklisted host.
Blacklisted the next available port and then added a new tserver on that port
and confirmed no load gets assigned to it.

Reviewers: kannan, karthik, bharat

Reviewed By: bharat

Subscribers: jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D194