Don't allow building with gcc on Mac OS X for now
Summary: It turned out /usr/bin/gcc was still invoking clang on Mac OS X. When I tried to use Homebrew-installed gcc, I ran into a bunch of errors related to the fact that the default standard C++ library implementation on Mac OS X is the newer library `libc++`, and third-party dependencies are being built with `libc++`, forcing us to use `libc++` with gcc because of the ABI incompatibility, but gcc uses the older library `libstdc++` by default and requires a bunch of custom configuration to build with `libc++` (http://libcxx.llvm.org/docs/UsingLibcxx.html#id4). Since Mac OS X builds are developer-only and we already build with both gcc and clang on Linux (although using libstdc++, as far as I understand), I thought it would be easier to not allow gcc builds on Mac OS X for now.

Test Plan:
- Run `yb_build.sh` on Mac OS X. Ensure it uses `clang`.
- Run `yb_build.sh --gcc` on Mac OS X. Ensure that it refuses to proceed.

Reviewers: bharat, bogdan

Subscribers: jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D187