Require a semicolon after test macros
Summary:
Require a semicolon after test macros by not providing it as part of macro definitions.
The whole purpose of the `do { ... } while (0)` macro idiom is to make the resulting macro
look like a function, and adding a semicolon in the macro definition allows omitting the semicolon
when the macro is being used, resulting in inconsistent code.

Also see http://stackoverflow.com/questions/1067226/c-multi-line-macro-do-while0-vs-scope-block
for details on the `do { ... } while (0)` macro idiom.

Test Plan: build the code

Reviewers: akashnil

Reviewed By: akashnil

Subscribers: jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D252