cuckoo_table_builder & cuckoo_table_read tests are meant to be run in debug mode only
Summary:
As indicated by the comments in CuckooHash():

  https://phabricator.dev.yugabyte.com/diffusion/YB/browse/master/src/rocksdb/table/cuckoo_table_factory.h;90526bed53860c1ddc3a1fb57dba3acd5cc7f6c6$22-26

the ability to pass in a get_slice_hash() callback function is meant for DEBUG mode tests only. But the Makefiles were setup to run the offending tests in release mode also.

Fixing the same, and also reverting the instrumentation I had added earlier to track this problem down.

Test Plan: unit tests

Reviewers: karthik, mikhail

Reviewed By: mikhail

Subscribers: jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D271