Fix hybrid_clock-test on Mac OS X
Summary:
WaitUntilAfter only returns a timeout if it actually has to wait, so we have to make sure the
time we give it to wait util is in the future. On Mac OS X we don't get any clock error info,
so we've been using 20usec clock error, which is too small (40usec easily passes between coming up
with the timestamp to wait until and actually invoking WaitUntilAfter). Bumping that minimum
error to 1ms (1000usec) in the test.

Test Plan: Run the test on Mac OS X in a loop 10000 times.

Reviewers: akashnil, kannan

Reviewed By: kannan

Subscribers: kannan

Differential Revision: https://phabricator.dev.yugabyte.com/D264