Fix load balancer segfault and normal leader picking
Summary:
Added failure mechanics for the rare case in which a Master Leader gets elected the very
first time and it starts doing load balancing while knowing all of the tablet related information,
but missing some of the tablet server information, as they have not yet successfully heartbeated,
in which case, you would end up with a null pointer access on the missing TSDescriptor.

Also fixed the fact that in normal conditions, we would not consider if we're moving a tablet away
from a tablet server that is actually its leader, in which case we could get stuck in a permanent
loop of adding a peer and removing the same peer right afterwards.

Test Plan: ran catalog_manager test, ran local cluster and added server and checked load.

Reviewers: karthik, mikhail, bharat

Reviewed By: bharat

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D286