ENG-227: use multiple drives when running unit tests
Summary:
Part 2 of changes to use multiple  drives.

The build-and-test scripts automatically detect if we are running in a mode where there are multiple drives mounted (at /mnt/ephemeral*). Each unit test randomly picks on of the drives as its log and data directory. Even in this mode, we leave symlinks so that the paths look similar to before. So the test logs and data directory can still be accessed from for example: $BUILD_ROOT/build/release-gcc-dynamic/yb-test-logs/<testName>/*.

Test Plan: Ran the scripts on my CentOS dev server. But need to test in Jenkins environment next in conjunction with the new AMI I have created for creating jenkins slaves with ephemeral drives.

Reviewers: mikhail

Reviewed By: mikhail

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D288