ENG-209: Adding wait for leader master task
Summary:
We should be explicitly waiting for a leader election to happen on the masters before we
issue commands such as changing the cluster config. Previously we were hacking around that with a
5 second wait. This makes it a more formal and explicit command.

This depends on https://phabricator.dev.yugabyte.com/D301.

Test Plan: created a cluster and saw the wait.

Reviewers: karthik, bharat

Reviewed By: bharat

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D302