Small fix on previously broken check in CatalogManager
Summary:
Had a small check problem in the code that was not manifesting in local testing as we
never actually really set the placement for the whole cluster. Not entirely sure how this didn't
get noticed through with the initial round of testing. We should look into adding more testing for
multi-placement deployments.

Test Plan:
changed code on devserver, pushed out, deployed cluster, used load tester from local
and wrote data to it.

Reviewers: karthik, bharat

Reviewed By: bharat

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D337