Fix RocksDB dynamic library dependency on Mac OS X
Summary:
Add a post-processing step to replace the "librocksdb_..." with "@rpath/librocksdb_..."
not just to unit tests, but also to all dynamic libraries and hand-picked set of executables
that are known to depend on RocksDB directly or indirectly. It is still not clear why this
is not happening during compilation time.

Test Plan: Unit tests

Reviewers: hector, akashnil

Reviewed By: akashnil

Subscribers: jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D358