[ENG-266] Fix intermittent failures of RaftConsensusITest.TestMasterNotifiedOnConfigChange with "Committed index is at: 2 and not yet at 1".
Summary:
RaftConsensusITest.TestMasterNotifiedOnConfigChange intermittently fails by repeatedly logging "Committed index is at: 2 and not yet at 1".
The issue is that there might be one or more split votes during the initial election, which could advance the index, so we can't guarantee
the index will always be 1. Instead, we grab the index that we've seen all servers agree on, and wait for the leader's committed index
to match that.

Also adding a script for running a specific test in a loop for the given number of times with the given level of parallelism.

Test failure:
https://detective.dev.yugabyte.com/job/yugabyte-centos-phabricator-clang-debug/47/artifact/build/debug-clang-dynamic/yb-test-logs/bin__raft_consensus-itest/RaftConsensusITest_TestMasterNotifiedOnConfigChange.log#line-375

Test Plan: - Ran the test 1000 times with 6-way parallelism on a 6-core machine with 100% success.

Reviewers: kannan

Reviewed By: kannan

Subscribers: jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D283