[ENG-50] Fix intermittent failures of RaftConsensusITest.TestReplicaBehaviorViaRPC
Summary:
The intermittent failures of RaftConsensusITest.TestReplicaBehaviorViaRPC seem to be caused by not
having enough time to send an RPC response after a server-side timeout. Trying to fix this by
ensuring that the server-side timeout will be at most 100 ms while the client-side timeout is 200 ms.

Test Plan: Run the test in a loop

Reviewers: kannan

Reviewed By: kannan

Subscribers: jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D368