[ENG-298] Unregister Memtrackers before their ownership is given up
Summary:
Unregister is not called on the root memtracker of log caches before it's destructor is called.

This causes dangling raw_pointers from its parent to itself for a brief period. Within that period, if a new log cache is created, it is confused.

Test Plan: Testing locally KVTableTest_Eng135MetricsTest with many iterations. Before the change, it fails once every 150 times in the local environment.

Reviewers: kannan, karthik, mikhail

Reviewed By: mikhail

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D407