ENG-351: Adding a UUID to the client and trickled down to the load tester so we can run them in parallel
Summary:
Added a UUID to the client and plumbed it down to the load tester and the keys. This way we can run multiple testers in parallel, with non-overlapping key spaces.

Also changed options in the tool, removed `create_table` and instead added `drop_table`. By default, a table will be created if it doesn't exists and appended to if it does exist. The table will only be dropped if the new flag is set!

Test Plan: Just run multiple `./bin/local_load_test.sh` in parallel, as first one will create the table and next ones will append.

Reviewers: bharat, karthik, kannan

Reviewed By: kannan

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D438